distribute functions in seprate modules, bug in Dijkastra
